+++ src/main/java/org/mmarini/wheelly/engines/FindLabelState.java
@@ -133,14 +133,16 @@ public record FindLabelState(String id, ProcessorCommand onInit, ProcessorComman
         }
         double growthDistance = getDouble(context, GROWTH_DISTANCE);
         Random random = get(context, RANDOM);
-        RRTDiscretePathFinder pathFinder = RRTDiscretePathFinder.createLabelTargets(map, robotLocation, distance, growthDistance, random,
-                Arrays.stream(labels));
+        RRTPathFinder pathFinder = RRTPathFinder.createLabelTargets(map, robotLocation, distance, growthDistance, random,
+                        Arrays.stream(labels))
+                .init();
 
         long timeout = System.currentTimeMillis() + getLong(context, MAX_SEARCH_TIME);
         // Look for the maximum time interval
         int minGoals = getInt(context, MIN_GOALS);
         int maxIterations = getInt(context, MAX_ITERATIONS);
         for (int i = 0; i < maxIterations
+                && !pathFinder.isCompleted()
                 && pathFinder.rrt().goals().size() < minGoals
                 && System.currentTimeMillis() <= timeout; i++) {
             pathFinder.grow();